requirements.txt
PySide6==6.7.0
psycopg2-binary==2.9.9
pandas==2.0.3
python-dotenv==1.0.0